Add and show can_disconnect status
Update model with new bool showing if or not the car can be disconnected. This commit includes the necessary logic changes to go along. Now websockets messages use JSON to be able to be parsed more precisely by JS. can_disconnect will only be shown to the System Engineer group in django admin. Changing its state will not cause a refresh on the client.